Fix app_ble.cfg for multirole & security
The examples' app_ble.cfg were hard-coded to 'c:/git/..' and that won't work for
users that clone their tree to a different place.

Fix:
Remove the hard-coded c:/git and have the .cfg file import starting from the
'src' directory in the BLE SDK.

For IAR:
The import will only work if the path to the BLE SDK is also included into
XDCPATH env variable. That change was done in *.custom_argsvar.

For CCS:
TBD